third_party/WebKit/LayoutTests/bluetooth/characteristic/getDescriptors/gen-descriptor-blocklisted.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  let promise;
  return setBluetoothFakeAdapter('DisconnectingHealthThermometerAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['health_thermometer']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService('health_thermometer'))
    .then(service => service.getCharacteristic('measurement_interval'))
    .then(characteristic => {
      return assert_promise_rejects_with_message(
        // Using UUIDs instead of names to avoid making a name<>uuid mapping.
        characteristic.getDescriptors('bad2ddcf-60db-45cd-bef9-fd72b153cf7c'),
          new DOMException('getDescriptor(s) called with blocklisted UUID. ' +
                         'https://goo.gl/4NeimX',
              'SecurityError'));
    })
}, 'Making sure getDescriptors can not access blocklisted descriptors.');

</script>
